requirements.txt
streamlit
pdfplumber
openai
pinecone-client
